refactor(frontend): organize sidebar's code
Moves the sidebar code into its own folder
