+++ frontend/src/App.tsx
@@ -21,7 +21,7 @@
 import { useState } from "react";
 import { Navigate, useRoutes } from "react-router-dom";
 
-import Sidebar from "@/components/Sidebar";
+import Sidebar from "@/components/sidebar/Sidebar";
 import Topbar from "@/components/Topbar";
 import { useAuth } from "@/contexts/Auth";
 import { Route } from "@/Navigation";
