+++ frontend/src/components/sidebar/Sidebar.scss
